Pull getImportResult to shared resolver
