+++ src/resolvers/import-resolvers.ts
@@ -10,16 +10,17 @@ import {
   ImportStatus,
   FeaturesList,
 } from '../resolverTypes';
-import {
-  clearImportResult,
-  getImportResult,
-  getImportStatus,
-  ImportFailedError,
-  importProjects,
-} from '../services/import-projects';
+import { clearImportResult, getImportStatus, importProjects } from '../services/import-projects';
 import { GroupProjectsResponse, TeamsWithMembershipStatus } from '../types';
 import { getAllComponentTypeIds } from '../client/compass';
-import { appId, connectedGroupsInfo, getFeatures, groupsAllExisting } from './shared-resolvers';
+import {
+  appId,
+  connectedGroupsInfo,
+  getFeatures,
+  getImportResult,
+  groupsAllExisting,
+  ImportFailedError,
+} from './shared-resolvers';
 import { getFirstPageOfTeamsWithMembershipStatus } from '../services/get-teams';
 import { getTeamOnboarding, setTeamOnboarding } from '../services/onboarding';
