+++ src/resolvers/shared-resolvers.ts
@@ -1,8 +1,25 @@
-import { AuthErrorTypes, DefaultErrorTypes, FeaturesList, GitlabAPIGroup, ResolverResponse } from '../resolverTypes';
+import { ListResult, startsWith, storage } from '@forge/api';
+import {
+  AuthErrorTypes,
+  DefaultErrorTypes,
+  FeaturesList,
+  GitlabAPIGroup,
+  ImportableProject,
+  ImportErrorTypes,
+  ProjectImportResult,
+  ResolverResponse,
+} from '../resolverTypes';
 import { listFeatures } from '../services/feature-flags';
 import { getAllExistingGroups, getConnectedGroups } from '../services/group';
 import { setupAndValidateWebhook } from '../services/webhooks';
 import { getForgeAppId } from '../utils/get-forge-app-id';
+import { STORAGE_KEYS } from '../constants';
+
+export class ImportFailedError extends Error {
+  constructor(readonly errorType: ImportErrorTypes, readonly message: string) {
+    super(message);
+  }
+}
 
 export const getFeatures = (): ResolverResponse<FeaturesList> => {
   try {
@@ -61,5 +78,25 @@ export const appId = (): ResolverResponse<string> => {
       success: false,
       errors: [{ message: e.message, errorType: DefaultErrorTypes.NO_APP_ID_VARIABLE_DEFINED }],
     };
+  }
+};
+
+export const getFailedProjects = (): Promise<ListResult> => {
+  const response = storage.query().where('key', startsWith(STORAGE_KEYS.CURRENT_IMPORT_FAILED_PROJECT_PREFIX));
+
+  return response.getMany();
+};
+
+export const getImportResult = async (): Promise<ProjectImportResult> => {
+  try {
+    const listFailedProjects = await getFailedProjects();
+    const failed = listFailedProjects.results.map(({ value }) => value as ImportableProject);
+    const total = await storage.get(STORAGE_KEYS.CURRENT_IMPORT_TOTAL_PROJECTS);
+    return {
+      failed,
+      total,
+    };
+  } catch (err) {
+    throw new ImportFailedError(ImportErrorTypes.CANNOT_GET_IMPORT_RESULT, err.message);
   }
 };
